requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
faiss-cpu
pydantic
pdfplumber
streamlit
requests
python-dotenv
gunicorn
